# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Passport Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List the products",
                "operationId": "list-products",
                "responses": {
                    "200": {
                        "description": "Products retrieved successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Product created successfully.",
                "operationId": "create-product",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "shortDesc",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product created successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/products/:id": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "operationId": "get-warehouses-by-id",
                "responses": {
                    "200": {
                        "description": "Producte retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "delete-warehouse",
                "responses": {
                    "200": {
                        "description": "Product deleted successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update a product",
                "operationId": "update-product",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "stock",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "shortDesc",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register an account",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User registered successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get access",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/warehouses": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "List the warehouses",
                "operationId": "list-warehouses",
                "responses": {
                    "200": {
                        "description": "Warehouses retrieved successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Add a warehouse",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse created successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Delete a warehouse",
                "operationId": "delete-warehouse",
                "responses": {
                    "200": {
                        "description": "Warehouses deleted successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/warehouses/:id": {
            "get": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Get warehouse by ID",
                "operationId": "get-warehouses-by-id",
                "responses": {
                    "200": {
                        "description": "Warehouse retrieved successfully",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Update a warehouse",
                "operationId": "update-warehouse",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse updated successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/attach": {
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Attach a product with a warehouse",
                "operationId": "attach-product-with-warehouse",
                "responses": {
                    "200": {
                        "description": "Products attached successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/detach": {
            "post": {
                "tags": [
                    "Warehouse"
                ],
                "summary": "Detach a product with a warehouse",
                "operationId": "detach-product-with-warehouse",
                "responses": {
                    "200": {
                        "description": "Products detached successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}